+++ package.json
@@ -20,7 +20,7 @@
     "node": "10"
   },
   "dependencies": {
-    "dayjs": "^1.9.5",
-    "flatpickr": "^4.6.6"
+    "dayjs": "1.9.5",
+    "flatpickr": "4.6.6"
   }
 }
