+++ src/const/const.js
@@ -1,0 +1,3 @@
+const EVENT_TYPES = [`taxi`, `bus`, `train`, `ship`, `transport`, `drive`, `flight`, `check-in`, `sightseeing`, `restaurant`];
+
+export {EVENT_TYPES};
