--- src/mock/filter.js
@@ -1,1 +1,0 @@
-
