+++ src/utils/event-type-to-uppercase.js
@@ -1,0 +1,4 @@
+const eventTypeToUpperCase = (type) => {
+  return type[0].toUpperCase() + type.slice(1);
+};
+export {eventTypeToUpperCase};
