--- src/view/event-edit.js
@@ -1,134 +1,0 @@
-import {EVENT_TYPES} from "../mock/point";
-
-const createEditPointTemplate = (point) => {
-  const {offers, destination, price} = point;
-
-  const renderTripTypesList = (types) => {
-    return types.map((type, id) => {
-      return (
-        `<div class="event__type-item">
-        <input id="event-type-${type.toLowerCase()}-${id + 1}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${type}">
-        <label class="event__type-label  event__type-label--${type.toLowerCase()}" for="event-type-${type.toLowerCase()}-${id + 1}">${type}</label>
-      </div>`.trim()
-      );
-    })
-      .join(`\n`);
-  };
-
-  const renderPhotos = (pictures) => {
-    return pictures.map((picture) => {
-      return (
-        `<img class="event__photo" src="${picture}" alt="Event photo">`.trim());
-    })
-      .join(`\n`);
-  };
-
-  const renderEventOffers = () => {
-    return offers.map((offer, index) => {
-      const {id, title} = offer;
-      const isChecked = Math.random() > 0.5;
-
-      return (
-        `<div class="event__offer-selector">
-        <input class="event__offer-checkbox  visually-hidden" id="event-offer-${id}-${index + 1}" type="checkbox" name="event-offer-${id}"
-        ${isChecked ? `checked` : ``}
-        >
-        <label class="event__offer-label" for="event-offer-${id}-${index + 1}">
-          <span class="event__offer-title">${title}</span>
-          &plus;
-          &euro;&nbsp;<span class="event__offer-price">${price}</span>
-        </label>
-      </div>`.trim()
-      );
-    })
-      .join(``);
-  };
-
-  const renderOffers = () => {
-    const eventOffers = renderEventOffers(offers);
-    return (
-      `<section class="event__section  event__section--offers">
-      <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-      <div class="event__available-offers">
-        ${eventOffers}
-      </div>
-    </section>`.trim()
-    );
-  };
-
-  const typesList = renderTripTypesList(EVENT_TYPES.slice(0, 7));
-  const eventPhotos = renderPhotos(destination.pictures);
-  const eventOffers = renderOffers(offers);
-
-  return `<li class="trip-events__item">
-             <form class="event event--edit" action="#" method="post">
-                <header class="event__header">
-                  <div class="event__type-wrapper">
-                    <label class="event__type  event__type-btn" for="event-type-toggle-1">
-                      <span class="visually-hidden">Choose event type</span>
-                      <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
-                    </label>
-                    <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
-
-                    <div class="event__type-list">
-                      <fieldset class="event__type-group">
-                        <legend class="visually-hidden">Event type</legend>
-
-                        <div class="event__type-item">
-                          ${typesList}
-                        </div>
-                      </fieldset>
-                    </div>
-                  </div>
-
-                  <div class="event__field-group  event__field-group--destination">
-                    <label class="event__label  event__type-output" for="event-destination-1">
-                      Flight
-                    </label>
-                    <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Geneva" list="destination-list-1">
-                    <datalist id="destination-list-1">
-                      <option value="Amsterdam"></option>
-                      <option value="Geneva"></option>
-                      <option value="Chamonix"></option>
-                    </datalist>
-                  </div>
-
-                  <div class="event__field-group  event__field-group--time">
-                    <label class="visually-hidden" for="event-start-time-1">From</label>
-                    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="19/03/19 00:00">
-                    &mdash;
-                    <label class="visually-hidden" for="event-end-time-1">To</label>
-                    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="19/03/19 00:00">
-                  </div>
-
-                  <div class="event__field-group  event__field-group--price">
-                    <label class="event__label" for="event-price-1">
-                      <span class="visually-hidden">Price</span>
-                      &euro;
-                    </label>
-                    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="${price}">
-                  </div>
-
-                  <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-                  <button class="event__reset-btn" type="reset">Cancel</button>
-                </header>
-                <section class="event__details">
-                  ${eventOffers}
-
-                  <section class="event__section  event__section--destination">
-                    <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-                    <p class="event__destination-description">${destination.description}</p>
-
-                    <div class="event__photos-container">
-                      <div class="event__photos-tape">
-                        ${eventPhotos}
-                      </div>
-                    </div>
-                  </section>
-                </section>
-              </form>
-             </li>
-`;
-};
-
-export {createEditPointTemplate};
