+++ src/view/event-edit/render-offers.js
@@ -1,0 +1,35 @@
+const renderEventOffers = (offers, price) => {
+
+  return offers.map((offer, index) => {
+    const {id, title} = offer;
+    const isChecked = Math.random() > 0.5;
+
+    return (
+      `<div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-${id}-${index + 1}" type="checkbox" name="event-offer-${id}"
+        ${isChecked ? `checked` : ``}
+        >
+        <label class="event__offer-label" for="event-offer-${id}-${index + 1}">
+          <span class="event__offer-title">${title}</span>
+          &plus;
+          &euro;&nbsp;<span class="event__offer-price">${price}</span>
+        </label>
+      </div>`.trim()
+    );
+  })
+    .join(``);
+};
+
+const renderOffers = (offers) => {
+  const eventOffers = renderEventOffers(offers);
+  return (
+    `<section class="event__section  event__section--offers">
+      <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+      <div class="event__available-offers">
+        ${eventOffers}
+      </div>
+    </section>`.trim()
+  );
+};
+
+export {renderOffers};
