+++ src/view/event-edit/render-options.js
@@ -1,0 +1,8 @@
+const renderOptions = (cities) => {
+  return cities.map((city) => {
+    return (`<option value="${city}"></option>`.trim());
+  })
+    .join(``);
+};
+
+export {renderOptions};
