+++ src/view/event-edit/render-photos.js
@@ -1,0 +1,9 @@
+const renderPhotos = (pictures) => {
+  return pictures.map((picture) => {
+    return (
+      `<img class="event__photo" src="${picture}" alt="Event photo">`);
+  })
+    .join(``);
+};
+
+export {renderPhotos};
