+++ src/view/event-item/event-item.js
@@ -1,0 +1,52 @@
+import {eventTypeToUpperCase} from "../../utils/event-type-to-uppercase";
+import {generateEventDate, generateEventStartTime, generateEventEndTime} from "../../utils/date";
+import {renderTripEventOffers} from "./render-offers-in-item";
+
+export const createEventItemTemplate = (point) => {
+  const {dateFrom, dateTo, type, destination, price, isFavorite, offers} = point;
+
+  const eventTypeUpperCase = eventTypeToUpperCase(type);
+  const eventDate = generateEventDate(dateFrom);
+  const eventStartTime = generateEventStartTime(dateFrom);
+  const eventEndTime = generateEventEndTime(dateTo);
+
+  const eventOffers = offers !== null ? renderTripEventOffers(offers) : ``;
+
+  const buttonFavoriteActive = isFavorite ? ` event__favorite-btn--active` : ``;
+
+  return `<li class="trip-events__item">
+            <div class="event">
+                <time class="event__date" datetime="${dateFrom.toISOString()}">${eventDate}</time>
+                <div class="event__type">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/${type}.png" alt="Event type icon">
+                </div>
+                <h3 class="event__title">${eventTypeUpperCase} to ${destination.name}</h3>
+                <div class="event__schedule">
+                  <p class="event__time">
+                    <time class="event__start-time" datetime="${dateFrom.toISOString()}}">${eventStartTime}</time>
+                    &mdash;
+                    <time class="event__end-time" datetime="${dateTo.toISOString()}">${eventEndTime}</time>
+                  </p>
+                  <p class="event__duration">1H</p>
+                </div>
+                <p class="event__price">
+                  &euro;&nbsp;<span class="event__price-value">${price}</span>
+                </p>
+
+                <h4 class="visually-hidden">Offers:</h4>
+                <ul class="event__selected-offers">
+                  ${eventOffers}
+                </ul>
+
+                <button class="event__favorite-btn ${buttonFavoriteActive}" type="button">
+                  <span class="visually-hidden">Add to favorite</span>
+                  <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
+                    <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
+                  </svg>
+                </button>
+                <button class="event__rollup-btn" type="button">
+                  <span class="visually-hidden">Open event</span>
+                </button>
+              </div>
+            </li>`;
+};
